Diplom_1/requirements.txt
allure-pytest==2.15.3
pytest==9.0.2
pytest-cov==7.1.0
